# backend/app/api/static/docs/openapi-3.yaml
openapi: 3.0.0
info:
  description: Track, Manage, and Organize your Things.
  title: HomeBot API
  contact:
    name: HomeBot Team
    url: https://github.com/pritish-codes/homebot
  version: "1.0"
paths:
...
  /v1/users/login/oidc/callback:
    get:
      tags:
        - Authentication
      summary: OIDC Callback Handler
      parameters:
        - description: Authorization code
          name: code
          in: query
          required: true
          schema:
            type: string
        - description: State parameter
          name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Found
...
servers:
  - url: /api
components:
  securitySchemes:
    Bearer:
      description: "\"Type 'Bearer TOKEN' to correctly set the API Key\""
      type: apiKey
      name: Authorization
      in: header
...